Use /v3 as prefix for /refresh
matrix-js-sdk previously used /v1 as prefix for /refresh endpoint. This
would work with Synapse v1.71.0 and below, as Synapse had a bug which
incorrectly used /v1 for /refresh. After matrix-org/synapse#14364 (which
was released in v1.72.0), using /v1 no longer works with Synapse.

This change will break matrix-js-sdk clients which are interfacing with
Synapse pre-v1.72.0; but using /v3 is the correct behavior.